+++ Hawk/functions/User/Start-HawkUserInvestigation.ps1
@@ -207,6 +207,20 @@
 					Out-LogFile "Running Get-HawkUserMailItemsAccessed" -Action
 					Get-HawkUserMailItemsAccessed -UserPrincipalName $User
 				}
+				if ($PSCmdlet.ShouldProcess("Running Get-HawkUserExchangeSearchQuery for $User")) {
+					Out-LogFile "Running Get-HawkUserExchangeSearchQuery" -Action
+					Get-HawkUserExchangeSearchQuery -UserPrincipalName $User
+				}
+
+				if ($PSCmdlet.ShouldProcess("Running Get-HawkUserMailSendActivity for $User")) {
+					Out-LogFile "Running Get-HawkUserMailSendActivity" -Action
+					Get-HawkUserMailSendActivity -UserPrincipalName $User
+				}
+
+				if ($PSCmdlet.ShouldProcess("Running Get-HawkUserSharePointSearchQuery for $User")) {
+					Out-LogFile "Running Get-HawkUserSharePointSearchQuery" -Action
+					Get-HawkUserSharePointSearchQuery -UserPrincipalName $User
+				}
 	
 				if ($PSCmdlet.ShouldProcess("Running Get-HawkUserMobileDevice for $User")) {
 					Out-LogFile "Running Get-HawkUserMobileDevice" -Action